requirements.txt
absl-py==0.11.0
astor==0.8.1
cached-property==1.5.2
click==7.1.2
cycler==0.10.0
dataclasses==0.7
fastapi==0.61.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.0.4
h11==0.11.0
h5py==3.1.0
importlib-metadata==2.0.0
imutils==0.5.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
numpy==1.19.4
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.1.4
Pillow==8.0.1
protobuf==3.14.0
pydantic==1.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2020.4
six==1.15.0
starlette==0.13.6
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
typing-extensions==3.7.4.3
uvicorn==0.12.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
